Exported databse for website
This is the database created for the website. So when making changes within wordpress that deal directly with text, images, etc, this will need to be exported/imported each time. In the future, in order to work on the wordpress site, import this database, then when finished, export it back to the repository